Fixed some more compiler warnings.

+++ licq/src/support.c
@@ -47,7 +47,7 @@ void SetString(char **szDest, const char *szSource)
 
 char *ParseDigits(char *szDest, const char *szSource, unsigned int nLen)
 {
-  int n = 0;
+  unsigned int n = 0;
   char *szCur = szDest;
 
   while ((*szSource) && (n < nLen))
@@ -109,7 +109,7 @@ int UinString(char *_szBuf, size_t _nSize, unsigned long _nUin)
   int nRet = snprintf(_szBuf, _nSize - 1, "%lu", _nUin);
   _szBuf[_nSize - 1] = '\0';
 
-  return (nRet == -1) ? _nSize - 1 : nRet; 
+  return (nRet == -1) ? (int)_nSize - 1 : nRet; 
 }
 
 int Redirect(const char *_szFile)
